% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.r
\name{generateDescriptiveStats}
\alias{generateDescriptiveStats}
\title{Generate descriptive statistics}
\usage{
generateDescriptiveStats(data, output_directory)
}
\arguments{
\item{data}{A data.table of prepared OHCAO data}

\item{output_directory}{Path to desired output directory,
a character vector of length one.}
}
\value{
Nothing. On success and html file named
"OHCAO_data_report.html" will be created in the
specified output directory.
}
\description{
Generate descriptive statistics
}
\examples{
\dontrun{
generateDescriptiveStats(ohcao_data, "path/to/output/dir/")
}
}
