% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ohcao_data.r
\name{prepareOHCAOData}
\alias{prepareOHCAOData}
\title{Prepare OHCAO data for analysis}
\usage{
prepareOHCAOData(ohcao_data_raw)
}
\arguments{
\item{ohcao_data_raw}{A data.table of raw OHCAO data.}
}
\value{
A data.table of prepared OHCAO data.
}
\description{
Prepare OHCAO data for analysis
}
\examples{
\dontrun{
prepareOHCAOData(ohcao_data_raw)
}
}
